[SYSTEMML-2366] Reorganise the test
